requirements.txt
Adafruit-Blinka==8.50.0
adafruit-blinka-displayio==0.11.1
adafruit-circuitpython-bitmap_font==2.1.3
adafruit-circuitpython-busdevice==5.2.10
adafruit-circuitpython-connectionmanager==3.1.2
adafruit-circuitpython-display-text==3.2.0
adafruit-circuitpython-display_button==1.11.1
adafruit-circuitpython-display_shapes==2.10.0
adafruit-circuitpython-requests==4.1.8
adafruit-circuitpython-ticks==1.1.1
adafruit-circuitpython-typing==1.11.2
Adafruit-PlatformDetect==3.75.0
Adafruit-PureIO==1.1.11
binho-host-adapter==0.1.6
blinka-displayio-pygamedisplay==2.4.0
numpy==2.2.0
pillow==11.0.0
pyftdi==0.56.0
pygame==2.6.1
pyserial==3.5
pyusb==1.2.1
typing_extensions==4.12.2
